@@ -5,7 +5,7 @@ publish_to: none
 # homepage: https://www.example.com
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
 
 dependencies:
   api_models:
