@@ -4,7 +4,7 @@ version: 0.0.28+31
 publish_to: none
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
