@@ -3,7 +3,7 @@ description: A starting point for Dart libraries or applications.
 version: 1.0.0
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
 
 
 dependencies:
