@@ -29,7 +29,7 @@ class ConfigurationView extends StatelessWidget {
           padding: const EdgeInsets.all(16),
           child: Text(
             'Configuration',
-            style: Theme.of(context).textTheme.headline5,
+            style: Theme.of(context).textTheme.headlineSmall,
           ),
         ),
         const VSpace(spacing: 16),
