@@ -19,7 +19,7 @@ class LogOutController extends StateNotifier<AsyncValue<void>> {
       // TODO(brandon): Is this the best way to do this?
       _ref.invalidate(authProvider);
     } on Exception catch (e) {
-      state = AsyncValue.error(e);
+      state = AsyncValue.error(e, StackTrace.current);
     } finally {
       state = const AsyncValue.data(null);
     }
